АПИ разделены по тематике. Произведена чистка кода.
