+++ src/components/ExchangeRates/exchangeRates.module.css
@@ -6,7 +6,6 @@
 }
 
 .currencySign {
-  /* color: #ffb780; */
   color: #ff9a57;
   margin-right: 0.3rem;
 }
